+++ FileReading/FileReader.hpp
@@ -43,54 +43,4 @@ public:
     };
 };
 
-//todo: make web reader
-
-// Helper function to handle data received from libcurl
-static size_t WriteCallback(void *contents, size_t size, size_t nmemb, std::string *output) {
-    size_t totalSize = size * nmemb;
-    output->append((char *) contents, totalSize);
-    return totalSize;
-}
-
-// WebReader class to read data from a URL
-class WebReader : public FileReader {
-public:
-    // std::vector<std::string> openFile(const std::string &url) override {
-    //     CURL *curl;
-    //     CURLcode res;
-    //     std::string readBuffer;
-    //
-    //     curl = curl_easy_init();
-    //     if (curl) {
-    //         curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-    //         curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-    //         curl_easy_setopt(curl, CURLOPT_WRITEDATA, &readBuffer);
-    //
-    //         // Perform the request, and get the response
-    //         res = curl_easy_perform(curl);
-    //         if (res != CURLE_OK) {
-    //             std::cerr << "curl_easy_perform() failed: " << curl_easy_strerror(res) << std::endl;
-    //         }
-    //
-    //         // Cleanup
-    //         curl_easy_cleanup(curl);
-    //     } else {
-    //         throw std::runtime_error("Failed to initialize CURL.");
-    //     }
-    //
-    //     // Parse the readBuffer into lines and return as vector of strings
-    //     std::vector<std::string> data;
-    //     std::istringstream stream(readBuffer);
-    //     std::string line;
-    //     while (std::getline(stream, line)) {
-    //         data.push_back(line);
-    //     }
-    //
-    //     return data;
-    // }
-};
-
-// DiskReader class (for future extension)
-class DiskReader : public FileReader {};
-
 #endif //FILEREADER_HPP
